+++ CHANGELOG.md
@@ -13,6 +13,9 @@ The format is based on the [KeepAChangeLog] project.
 - [#847] Using pydantic for settings instead of custom class
 - [#851], [#852] Add `authn_method` to `Consumer.complete`
 
+## Fixed
+- [#XXX] Made oauth_example less broken
+
 [#847]: https://github.com/CZ-NIC/pyoidc/pull/847
 [#851]: https://github.com/CZ-NIC/pyoidc/issues/851
 [#852]: https://github.com/CZ-NIC/pyoidc/pull/852